+++ docs/CHANGELOG.md
@@ -3,6 +3,7 @@
 * Sep 2020
 	* FM radio project  added
 
-* April 2020
+* April 2021
 	* Tm1638plus library added 
 	* ERM19264_UC1609C library added
+	* Nokia 5110 text library added